ValidationMethodsTraitおよびFilterMethodsTraitを追加し、cfFormMailerクラスにバリデーションとフィルターメソッドを統合
