+++ includes/class.cfFormMailer.inc.php
@@ -10,8 +10,14 @@
  * LICENSE: GNU General Public License (GPL) (http://www.gnu.org/copyleft/gpl.html)
  */
 
+// トレイトの読み込み
+require_once __DIR__ . '/traits/ValidationMethodsTrait.php';
+require_once __DIR__ . '/traits/FilterMethodsTrait.php';
+
 class Class_cfFormMailer
 {
+    use ValidationMethodsTrait;
+    use FilterMethodsTrait;
 
     public $cfg = array();
 
@@ -2124,450 +2130,5 @@ class Class_cfFormMailer
             db()->config['dbase']
         );
         return db()->getRecordCount(db()->query($sql)) == 2;
-    }
-
-    /* ------------------------------------------------------------------ */
-    /* 標準装備の検証メソッド
-    /* ------------------------------------------------------------------ */
-
-    /**
-     * num : 数値？
-     */
-    private function _def_num($value, $param, $field)
-    {
-        // 強制的に半角に変換します。
-        $this->form[$field] = mb_convert_kana(
-            $this->form[$field] ?? '',
-            'n',
-            $this->config('charset')
-        );
-
-        if (is_numeric($this->form[$field])) {
-            return true;
-        }
-
-        return '半角数字で入力してください';
-    }
-
-    /**
-     * email : 正しいメールアドレス形式か？
-     */
-    private function _def_email($value, $param, $field)
-    {
-        // 強制的に半角に変換します。
-        $this->form[$field] = mb_convert_kana(
-            $this->form[$field] ?? '',
-            'a',
-            $this->config('charset')
-        );
-
-        if ($this->_isValidEmail($this->form[$field])) {
-            return true;
-        }
-
-        return 'メールアドレスの形式が正しくありません';
-    }
-
-    /**
-     * len(min, max) : 文字数チェック
-     */
-    private function _def_len($value, $param, $field)
-    {
-
-        if (!preg_match("/([0-9]+)?(-)?([0-9]+)?/", $param, $match)) {
-            return true;
-        }
-
-        if ($match[1] && empty($match[2]) && empty($match[3])) {
-            if (mb_strlen($value) != $match[1]) {
-                return $match[1] . '文字で入力してください';
-            }
-            return true;
-        }
-
-        if (!$match[1] && $match[2] && $match[3]) {
-            if (mb_strlen($value) > $match[3]) {
-                return $match[3] . '文字以内で入力してください';
-            }
-            return true;
-        }
-
-        if ($match[1] && $match[2] && !$match[3]) {
-            if (mb_strlen($value) < $match[1]) {
-                return $match[1] . '文字以上で入力してください';
-            }
-            return true;
-        }
-
-        if ($match[1] && $match[2] && $match[3]) {
-            if (mb_strlen($value) < $match[1] || mb_strlen($value) > $match[3]) {
-                return sprintf('%s～%s文字で入力してください', $match[1], $match[3]);
-            }
-            return true;
-        }
-        return true;
-    }
-
-    /**
-     * vericode : CAPTCHA による認証チェック
-     *   Added in v0.0.5
-     */
-    private function _def_vericode($value, $param, $field)
-    {
-        if (!$this->config('vericode')) {
-            return true;
-        }
-        if (sessionv('veriword') == $value) {
-            return true;
-        }
-
-        $this->form[$field] = '';
-        return '入力値が正しくありません';
-    }
-
-    /**
-     * range(min, max) : 値範囲チェック
-     *   Added in v0.0.5
-     */
-    private function _def_range($value, $param, $field)
-    {
-
-        if (!preg_match('/([0-9-]+)?(~)?([0-9-]+)?/', $param, $match)) {
-            return true;
-        }
-
-        if ($match[1] && !$match[2] && !$match[3]) {
-            if ($match[1] < $value) {
-                return '入力値が範囲外です';
-            }
-
-            return true;
-        }
-
-        if (!$match[1] && $match[2] && $match[3]) {
-            if ($match[3] < $value) {
-                return '入力値が範囲外です';
-            }
-            return true;
-        }
-
-        if (isset($match[1], $match[2])) {
-            if ($value < $match[1]) {
-                return '入力値が範囲外です';
-            }
-
-            if ($match[3] < $value) {
-                return '入力値が範囲外です';
-            }
-            return true;
-        }
-
-        return true;
-    }
-
-    /**
-     * sameas(field) : 同一確認
-     *   Added in v0.0.6
-     */
-    private function _def_sameas($value, $param, $field)
-    {
-        if ($value == ($this->form[$param] ?? null)) {
-            return true;
-        }
-
-        unset($this->form[$field]);
-        return sprintf(
-            '&laquo; %s &raquo; と一致しません',
-            $this->adaptEncoding(
-                $this->parsedForm[$param]['label'] ?? $param
-            )
-        );
-    }
-
-    /**
-     * tel : 電話番号？
-     *   Added in v0.0.7
-     */
-    private function _def_tel($value, $param, $field)
-    {
-        // 強制的に半角に変換します。
-        $this->form[$field] = mb_convert_kana($this->form[$field] ?? '', 'a', $this->config('charset'));
-        $this->form[$field] = preg_replace('@([0-9])ー@u', '$1-', $this->form[$field]);
-        if ((strpos($this->form[$field], '0') === 0)) {
-            $checkLen = 10;
-        } else {
-            $checkLen = 5;
-        }
-        $checkStr = preg_replace('/[^0-9]/', '', $this->form[$field]);
-        if (!preg_match('/[0-9]{4}$/', $this->form[$field])) {
-            return '正しい電話番号を入力してください。';
-        }
-
-        if (preg_match("/[^0-9\-+]/", $checkStr) || strlen($checkStr) < $checkLen) {
-            return '半角数字とハイフンで正しく入力してください';
-        }
-
-        return true;
-    }
-
-    /**
-     * tel : 郵便番号
-     *   Added in v1.3.x
-     */
-    private function _def_zip($value, $param, $field)
-    {
-        // 強制的に半角に変換します。
-        $this->form[$field] = mb_convert_kana($this->form[$field] ?? '', 'as', $this->config('charset'));
-        $this->form[$field] = preg_replace('/[^0-9]/', '', $this->form[$field]);
-        $str = $this->form[$field];
-
-        if (strlen($str) !== 7) {
-            return '半角数字とハイフンで正しく入力してください';
-        }
-
-        $this->form[$field] = substr($str, 0, 3) . '-' . substr($str, -4);
-
-        return true;
-    }
-
-    /**
-     * allowtype(type) : アップロードを許可するファイル形式
-     *   Added in v1.0
-     */
-    private function _def_allowtype($value, $param, $field)
-    {
-        $file_tmp_name = $_FILES[$field]['tmp_name'] ?? '';
-        if (empty($file_tmp_name) || !is_uploaded_file($file_tmp_name)) {
-            return true;
-        }
-        $allow_list = explode('|', $param);
-        if (!count($allow_list)) {
-            return true;
-        }
-
-        $mime = $this->_getMimeType($file_tmp_name, $field);
-        if ($mime === false) {
-            return '許可されたファイル形式ではありません';
-        }
-
-        $type = $this->_getType($mime);
-        if (!$type || !in_array($type, $allow_list, true)) {
-            return '許可されたファイル形式ではありません';
-        }
-
-        return true;
-    }
-
-    private function array_get($array, $key, $default = null)
-    {
-        if (!isset($array[$key])) {
-            return $default;
-        }
-        return $array[$key];
-    }
-    private function isSpam()
-    {
-        if (isset($_SERVER['HTTP_ACCEPT_LANGUAGE'])) {
-            return false;
-        }
-        return true;
-    }
-    /**
-     * allowsize(size) : アップロードを許可する最大ファイルサイズ
-     *   Added in v1.0
-     */
-    private function _def_allowsize($value, $param, $field)
-    {
-        $file_tmp_name = $_FILES[$field]['tmp_name'] ?? '';
-        if (empty($file_tmp_name) || !is_uploaded_file($file_tmp_name)) {
-            return true;
-        }
-
-        if (!$param || !is_numeric($param)) {
-            return false;
-        }
-
-        $size = @stat($file_tmp_name);
-        if ($size === false) {
-            return 'ファイルのアップロードに失敗しました';
-        }
-
-        if (($size['size'] <= $param * 1024)) {
-            return true;
-        }
-
-        return $param . 'キロバイト以内のファイルを指定してください';
-    }
-
-    /**
-     * convert(param)：半角英数字に変換
-     *   See: http://jp2.php.net/manual/ja/function.mb-convert-kana.php
-     *   Added in v1.2
-     * @param $value
-     * @param string $param
-     * @param $field
-     * @return bool
-     */
-    private function _def_convert($value, $param = 'K', $field)
-    {
-        if (!$param) {
-            $param = 'K';
-        }
-        $this->form[$field] = mb_convert_kana(
-            $this->form[$field] ?? '',
-            $param,
-            $this->config('charset')
-        );
-        return true;  // 常にtrueを返す
-    }
-
-    /**
-     * zenhan：半角英数字記号に変換
-     *   See: http://jp2.php.net/manual/ja/function.mb-convert-kana.php
-     *   Added in v1.2
-     * @param $value
-     * @param string $param
-     * @param $field
-     * @return bool
-     */
-    private function _def_zenhan($value, $param = 'VKas', $field)
-    {
-        $this->form[$field] = mb_convert_kana(
-            $this->form[$field] ?? '',
-            $param,
-            $this->config('charset')
-        );
-        $this->form[$field] = preg_replace('@([0-9])ー@u', '$1-', $this->form[$field]);
-        return true;  // 常にtrueを返す
-    }
-
-    /**
-     * hanzen：全角英数字記号に変換
-     *   See: http://jp2.php.net/manual/ja/function.mb-convert-kana.php
-     *   Added in v1.2
-     * @param $value
-     * @param string $param
-     * @param $field
-     * @return bool
-     */
-    private function _def_hanzen($value, $param = 'VKAS', $field)
-    {
-        $this->form[$field] = mb_convert_kana(
-            $this->form[$field] ?? '',
-            $param,
-            $this->config('charset')
-        );
-        return true;  // 常にtrueを返す
-    }
-
-    /**
-     * url(string)：URL値検証
-     *   Added in v1.2
-     */
-    private function _def_url($value, $param, $field)
-    {
-        return preg_match("@^https?://.+\..+@", $value);
-    }
-
-    /* ------------------------------------------------------------------ */
-    /* 標準装備のフィルターメソッド
-    /* ------------------------------------------------------------------ */
-
-    /**
-     * implode : 文字列結合
-     */
-    private function _f_implode($text, $param)
-    {
-        if (is_array($text)) {
-            return implode(str_replace("\\n", "\n", $param), $text);
-        }
-
-        return $text;
-    }
-
-    /**
-     * implodetag(tag) : HTMLタグで文字列結合
-     */
-    private function _f_implodetag($text, $param)
-    {
-
-        if (!is_array($text)) {
-            $text = array($text);
-        }
-
-        $ret = '';
-        foreach ($text as $v) {
-            $ret .= sprintf('<%s>%s</%s>', $param, $v, $param);
-        }
-        return $ret;
-    }
-
-    /**
-     * num : 数値のフォーマット （※PHP関数 number_format() と同様）
-     */
-    private function _f_num($text, $param)
-    {
-
-        if (is_array($text)) {
-            return array_map([$this, '_f_num'], $text, $param);
-        }
-
-        return number_format($text);
-    }
-
-    /**
-     * dateformat(format) : 日付のフォーマット
-     *
-     * strftime形式（%Y-%m-%d等）とdate形式（Y-m-d等）の両方に対応
-     */
-    private function _f_dateformat($text, $param)
-    {
-        if (is_array($text)) {
-            return array_map([$this, '_f_dateformat'], $text, $param);
-        }
-
-        $timestamp = strtotime($text);
-        if ($timestamp === false) {
-            return $text;
-        }
-
-        $datetime = (new DateTime())->setTimestamp($timestamp);
-
-        // strftimeフォーマット（%で始まる）が含まれている場合は変換
-        if (strpos($param, '%') !== false) {
-            $format_map = [
-                '%Y' => 'Y',  // 4桁の年
-                '%y' => 'y',  // 2桁の年
-                '%m' => 'm',  // 月（01-12）
-                '%d' => 'd',  // 日（01-31）
-                '%H' => 'H',  // 時（00-23）
-                '%M' => 'i',  // 分（00-59）
-                '%S' => 's',  // 秒（00-59）
-                '%B' => 'F',  // 月の完全名
-                '%b' => 'M',  // 月の短縮名
-                '%A' => 'l',  // 曜日の完全名
-                '%a' => 'D',  // 曜日の短縮名
-                '%e' => 'j',  // 日（1-31）
-                '%I' => 'h',  // 12時間制の時（01-12）
-                '%p' => 'A',  // AM/PM
-                '%w' => 'w',  // 曜日（0-6）
-            ];
-            $param = str_replace(array_keys($format_map), array_values($format_map), $param);
-        }
-
-        return $datetime->format($param);
-    }
-
-    /**
-     * sprintf(format) : テキストのフォーマット （※PHP関数 sprintf() と同様）
-     */
-    private function _f_sprintf($text, $param)
-    {
-
-        if (is_array($text)) {
-            return array_map([$this, '_f_sprintf'], $text, $param);
-        }
-
-        return sprintf($param, $text);
     }
 }
